+++ src/type_pool.cpp
@@ -101,6 +101,7 @@ TypePool::TypePool() {
   this->initErrorType(TYPE::StackOverflowError, "StackOverflowError");
   this->initErrorType(TYPE::RegexSyntaxError, "RegexSyntaxError");
   this->initErrorType(TYPE::RegexMatchError, "RegexMatchError");
+  this->initErrorType(TYPE::TildeError, "TildeError");
   this->initErrorType(TYPE::GlobbingError, "GlobbingError");
   this->initErrorType(TYPE::UnwrappingError, "UnwrappingError");
   this->initErrorType(TYPE::IllegalAccessError, "IllegalAccessError");